Inline ResponseCommon
This is an artifact of our multiplatform exploration.
