+++ okhttp/src/commonJvmAndroid/kotlin/okhttp3/Response.kt
@@ -17,29 +17,16 @@ package okhttp3
 
 import java.io.Closeable
 import java.io.IOException
+import java.net.HttpURLConnection.HTTP_MOVED_PERM
+import java.net.HttpURLConnection.HTTP_MOVED_TEMP
+import java.net.HttpURLConnection.HTTP_MULT_CHOICE
 import java.net.HttpURLConnection.HTTP_PROXY_AUTH
+import java.net.HttpURLConnection.HTTP_SEE_OTHER
 import java.net.HttpURLConnection.HTTP_UNAUTHORIZED
 import okhttp3.ResponseBody.Companion.asResponseBody
-import okhttp3.internal.commonAddHeader
-import okhttp3.internal.commonBody
-import okhttp3.internal.commonCacheControl
-import okhttp3.internal.commonCacheResponse
-import okhttp3.internal.commonClose
-import okhttp3.internal.commonCode
-import okhttp3.internal.commonHeader
-import okhttp3.internal.commonHeaders
-import okhttp3.internal.commonIsRedirect
-import okhttp3.internal.commonIsSuccessful
-import okhttp3.internal.commonMessage
-import okhttp3.internal.commonNetworkResponse
-import okhttp3.internal.commonNewBuilder
-import okhttp3.internal.commonPriorResponse
-import okhttp3.internal.commonProtocol
-import okhttp3.internal.commonRemoveHeader
-import okhttp3.internal.commonRequest
-import okhttp3.internal.commonToString
-import okhttp3.internal.commonTrailers
 import okhttp3.internal.connection.Exchange
+import okhttp3.internal.http.HTTP_PERM_REDIRECT
+import okhttp3.internal.http.HTTP_TEMP_REDIRECT
 import okhttp3.internal.http.parseChallenges
 import okio.Buffer
 
@@ -154,7 +141,7 @@ class Response internal constructor(
    * Returns true if the code is in [200..300), which means the request was successfully received,
    * understood, and accepted.
    */
-  val isSuccessful: Boolean = commonIsSuccessful
+  val isSuccessful: Boolean = code in 200..299
 
   @JvmName("-deprecated_message")
   @Deprecated(
@@ -172,13 +159,13 @@ class Response internal constructor(
   )
   fun handshake(): Handshake? = handshake
 
-  fun headers(name: String): List<String> = commonHeaders(name)
+  fun headers(name: String): List<String> = headers.values(name)
 
   @JvmOverloads
   fun header(
     name: String,
     defaultValue: String? = null,
-  ): String? = commonHeader(name, defaultValue)
+  ): String? = headers[name] ?: defaultValue
 
   @JvmName("-deprecated_headers")
   @Deprecated(
@@ -223,10 +210,14 @@ class Response internal constructor(
   )
   fun body() = body
 
-  fun newBuilder(): Builder = commonNewBuilder()
+  fun newBuilder(): Builder = Builder(this)
 
   /** Returns true if this response redirects to another resource. */
-  val isRedirect: Boolean = commonIsRedirect
+  val isRedirect: Boolean =
+    when (code) {
+      HTTP_PERM_REDIRECT, HTTP_TEMP_REDIRECT, HTTP_MULT_CHOICE, HTTP_MOVED_PERM, HTTP_MOVED_TEMP, HTTP_SEE_OTHER -> true
+      else -> false
+    }
 
   @JvmName("-deprecated_networkResponse")
   @Deprecated(
@@ -279,7 +270,14 @@ class Response internal constructor(
    */
   @get:JvmName("cacheControl")
   val cacheControl: CacheControl
-    get() = commonCacheControl
+    get() {
+      var result = lazyCacheControl
+      if (result == null) {
+        result = CacheControl.parse(headers)
+        lazyCacheControl = result
+      }
+      return result
+    }
 
   @JvmName("-deprecated_cacheControl")
   @Deprecated(
@@ -311,9 +309,11 @@ class Response internal constructor(
    * Prior to OkHttp 5.0, it was an error to close a response that is not eligible for a body. This
    * includes the responses returned from [cacheResponse], [networkResponse], and [priorResponse].
    */
-  override fun close() = commonClose()
+  override fun close() {
+    body.close()
+  }
 
-  override fun toString(): String = commonToString()
+  override fun toString(): String = "Response{protocol=$protocol, code=$code, message=$message, url=${request.url}}"
 
   open class Builder {
     internal var request: Request? = null
@@ -352,13 +352,25 @@ class Response internal constructor(
       this.trailersFn = response.trailersFn
     }
 
-    open fun request(request: Request) = commonRequest(request)
+    open fun request(request: Request) =
+      apply {
+        this.request = request
+      }
 
-    open fun protocol(protocol: Protocol) = commonProtocol(protocol)
+    open fun protocol(protocol: Protocol) =
+      apply {
+        this.protocol = protocol
+      }
 
-    open fun code(code: Int) = commonCode(code)
+    open fun code(code: Int) =
+      apply {
+        this.code = code
+      }
 
-    open fun message(message: String) = commonMessage(message)
+    open fun message(message: String) =
+      apply {
+        this.message = message
+      }
 
     open fun handshake(handshake: Handshake?) =
       apply {
@@ -372,7 +384,9 @@ class Response internal constructor(
     open fun header(
       name: String,
       value: String,
-    ) = commonHeader(name, value)
+    ) = apply {
+      headers[name] = value
+    }
 
     /**
      * Adds a header with [name] to [value]. Prefer this method for multiply-valued
@@ -381,24 +395,60 @@ class Response internal constructor(
     open fun addHeader(
       name: String,
       value: String,
-    ) = commonAddHeader(name, value)
+    ) = apply {
+      headers.add(name, value)
+    }
 
     /** Removes all headers named [name] on this builder. */
-    open fun removeHeader(name: String) = commonRemoveHeader(name)
+    open fun removeHeader(name: String) =
+      apply {
+        headers.removeAll(name)
+      }
 
     /** Removes all headers on this builder and adds [headers]. */
-    open fun headers(headers: Headers) = commonHeaders(headers)
+    open fun headers(headers: Headers) =
+      apply {
+        this.headers = headers.newBuilder()
+      }
 
-    open fun body(body: ResponseBody) = commonBody(body)
+    open fun body(body: ResponseBody) =
+      apply {
+        this.body = body
+      }
 
-    open fun networkResponse(networkResponse: Response?) = commonNetworkResponse(networkResponse)
+    open fun networkResponse(networkResponse: Response?) =
+      apply {
+        checkSupportResponse("networkResponse", networkResponse)
+        this.networkResponse = networkResponse
+      }
 
-    open fun cacheResponse(cacheResponse: Response?) = commonCacheResponse(cacheResponse)
+    open fun cacheResponse(cacheResponse: Response?) =
+      apply {
+        checkSupportResponse("cacheResponse", cacheResponse)
+        this.cacheResponse = cacheResponse
+      }
 
-    open fun priorResponse(priorResponse: Response?) = commonPriorResponse(priorResponse)
+    private fun checkSupportResponse(
+      name: String,
+      response: Response?,
+    ) {
+      response?.apply {
+        require(networkResponse == null) { "$name.networkResponse != null" }
+        require(cacheResponse == null) { "$name.cacheResponse != null" }
+        require(priorResponse == null) { "$name.priorResponse != null" }
+      }
+    }
+
+    open fun priorResponse(priorResponse: Response?) =
+      apply {
+        this.priorResponse = priorResponse
+      }
 
     @ExperimentalOkHttpApi
-    open fun trailers(trailersFn: (() -> Headers)): Builder = commonTrailers(trailersFn)
+    open fun trailers(trailersFn: (() -> Headers)): Builder =
+      apply {
+        this.trailersFn = trailersFn
+      }
 
     open fun sentRequestAtMillis(sentRequestAtMillis: Long) =
       apply {
